requirements.txt
Flask
flask_bcrypt
requests
pyzbar
pillow